Taylor Truncation

Taylor Truncation

Taylor truncation based Hamiltonian simulation (https://arxiv.org/abs/1412.4687) has many clear advantages. It has better complexity dependence on the precision and allows a greater range of Hamiltonians to be simulated.

The package provides circuits for five kinds of problems:

Unitary Taylor simulation

Non-unitary Taylor simulation

Unitary QuLDE Problem

Non-unitary QuLDEProblem

Solution by linearising a non-linear differential equation

QuDiffEq.TaylorParam — Type.

TaylorParam(k::Int, t::L, H::Matrix, x::Array{CPType,1})

Assigns values to parameters required for Taylor series based Hamiltonian simulation.

* k : sets order of Taylor series expansion
* t : time of evolution
* H : Hamiltonian
* x : intial vector

TaylorParam(k::Int,t::L,prob::QuLDEProblem{uType, tType, isinplace, F, P, T})

* k : sets order of Taylor series expansion
* t : time of evolution
* prob : wrapper for Linear differential equation problems (contains H, x, b)

source

QuDiffEq.circuit_ends — Method.

circuit_ends(n::Int, blk::TaylorParam{CPType, false}, VS1::AbstractMatrix, VS2::AbstractMatrix, VT::AbstractMatrix)

Generates the part of circuit that computes and decomputes the superposition of the ancilla bits, in non-unitary H quldecircuit

source

QuDiffEq.circuit_ends — Method.

circuit_ends(n::Int, blk::TaylorParam{CPType, false}, VS1::AbstractMatrix, VT::AbstractMatrix)

Generates the part of circuit that computes and decomputes the superposition of the ancilla bits, in non-unitary H taylorcircuit

source

QuDiffEq.circuit_ends — Method.

circuit_ends(n::Int, blk::TaylorParam{CPType, true}, VS1::AbstractMatrix, VS2::AbstractMatrix)

Generates the part of circuit that computes and decomputes the superposition of the ancilla bits, in unitary H quldecircuit

source

QuDiffEq.circuit_ends — Method.

circuit_ends(n::Int, blk::TaylorParam{CPType, true}, VS1::AbstractMatrix)

Generates the part of circuit that computes and decomputes the superposition of the ancilla bits, in unitary H taylorcircuit

source

QuDiffEq.circuit_intermediate — Method.

circuit_intermediate(n::Int, c::Int, blk::TaylorParam{CPType, false})

Generates the intermediate part of the circuit for non-unitary H.

source

QuDiffEq.circuit_intermediate — Method.

circuit_intermediate(n::Int, c::Int, blk::TaylorParam{CPType, true})

Generates the intermediate part of the circuit for unitary H.

source

QuDiffEq.taylorcircuit — Method.

taylorcircuit(n::Int, blk::TaylorParam, VS1::Matrix, VT::Matrix) ->->  ChainBlock{n}

Generates circuit for a non-unitary H input.

source

QuDiffEq.taylorcircuit — Method.

taylorcircuit(n::Int, blk::TaylorParam, VS1::Matrix) ->  ChainBlock{n}

Generates circuit for a unitary H input.

source

QuDiffEq.taylorsolve — Method.

taylorsolve(H::Array{CPType,2}, x::Vector{CPType}, k::Int, t::Real) -> ArrayReg, CPType

Simulates a Hamiltonian using the Taylor truncation method. Returns the state register and inverse probability of finding it.

source

QuDiffEq.v — Method.

v(n::Int,c::Int, T::Int, V::AbstractMatrix) -> ChainBlock{n}

Builds T input block. n : total number of qubits V : T input matrix c : starting qubit

source

QuDiffEq.v — Method.

v(n::Int,c::Int, j::Tuple, T::Int, V::AbstractMatrix) -> ControlBlock{n}

Builds a T input, j - control block. n : total number of qubits V : T input matrix c : starting qubit j : control bits tuple

source

QuDiffEq.calc_vs1 — Method.

calc_vs1(blk::TaylorParam, x::Vector{CPType}, opn::Real) ->  Matrix{CPType}

Calculates VS1 block for Taylor circuit.

source

QuDiffEq.calc_vs2 — Method.

calc_vs2(blk::TaylorParam, x::Vector{CPType}, opn::Real) ->  Matrix{CPType}

opn: operator norm of the input matrix.

Calculates VS2 block for Taylor circuit.

source

QuDiffEq.calc_vt — Method.

calc_vt(T)

calc_vt(::Type{CPType}) -> Matrix{CPType}

Generates VT block for non-unitary Taylor circuit.

source

QuDiffEq.unitary_decompose — Method.

unitary_decompose(H::Array{T,2}) -> Array{Array{T,2},1}

Generates a linear compostion of unitary matrices for argument H.

source

Quantum linear differential equation solver

A linear differential equation has the form :

\[ \frac{dx}{dt} = Mx + b\]

$M$is an arbitrary $N × N$ matrix,$x$and $b$are dimensional vectors. The exact solution for $x(t)$ is give by -

\[ x(t) = e^{Mt}x(0) + (e^{Mt} - I)M^{-1}b\]

This can be Taylor expanded up to the $k^{th}$order as -

\[x(t) \approx \sum^{k}_{m=0}\frac{(Mt)^{m}}{m!}x(0) + \sum^{k-1}_{n=1}\frac{(Mt)^{n-1}t}{n!}b\]

The vectors $x(0)$ and $b$ are encoded as state - $|x(0)\rangle = \sum_{i} \frac{x_{i}}{||x||} |i\rangle$ and $|b\rangle = \sum_{i} \frac{b_{i}}{||b||} |i\rangle$ with $\{|i \rangle\}$as the computational basis states. We can also write $M$ as $M = ||M||\mathcal{M}$. We then get:

\[|x(t)\rangle \approx \sum^{k}_{m=0}\frac{||x(0)||(||M||t)^{m}}{m!}\mathcal{M}^{m}|x(0)\rangle + \sum^{k-1}_{n=1}\frac{||b||(||M||t)^{n-1}t}{n!}\mathcal{M}^{n-1}|b\rangle\]

The quldecircuit effects this transformation on the input state to obtain $x(t)$.

QuDiffEq.quldecircuit — Method.

quldecircuit(n::Int,blk::TaylorParam,VS1::AbstractMatrix,VS2::AbstractMatrix,VT::AbstractMatrix) -> ChainBlock{n}

Generates circuit for solving linear differental equations for a non-unitary H input.

source

QuDiffEq.quldecircuit — Method.

quldecircuit(n::Int,blk::TaylorParam,VS1::AbstractMatrix,VS2::AbstractMatrix) -> ChainBlock{n}

Generates circuit for solving linear differental equations for a unitary H input.

source

Quantum non-linear differential equation

The non-linear solver constitutes two sub-routines. Firstly, the function transform sub-routine, which employs of the taylorcircuit. The function transform lets us map $z$ to $P(z)$, where $P$ is a quadratic polynomial. Secondly, the forward Euler method.

QuDiffEq.euler_matrix — Method.

euler_matrix(A::Matrix{CPType},b::Vector{CPType},h::Real) -> Matrix

Generates matrix for forward Euler iteration.

source

QuDiffEq.euler_matrix_update — Method.

euler_matrix(A::Matrix{CPType},b::Vector{CPType},h::Real) -> Matrix

Updates euler matrix for forward Euler iteration.

source

QuDiffEq.func_transform — Function.

func_transform(A::Matrix, x::Vector, k::Int) -> ArrayReg, <: Complex

Function transform sub-routine. Returns state register and inverse probability of finding it.

source

QuDiffEq.make_hermitian — Method.

make_hermitian(A::Matrix) -> Matrix

Returns hermitian matrix containing A.

source

QuDiffEq.make_input_vector — Method.

make_input_vector(x::Vector{T}) -> Vector{T}

Generates input vector for QuNLDE iterations.

source